Rename abstract class and interface names
