+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/source/AbstractSourceInfo.java
@@ -1,0 +1,17 @@
+package com.microsoft.azure.kusto.ingest.source;
+
+import java.util.UUID;
+
+abstract class AbstractSourceInfo implements SourceInfo {
+
+    private UUID sourceId;
+
+    public UUID getSourceId() {
+        return sourceId;
+    }
+
+    public void setSourceId(UUID sourceId) {
+        this.sourceId = sourceId;
+    }
+
+}
