--- ingest/src/main/java/com/microsoft/azure/kusto/ingest/source/ISourceInfo.java
@@ -1,5 +1,0 @@
-package com.microsoft.azure.kusto.ingest.source;
-
-public interface ISourceInfo {
-    void validate();
-}
