+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/source/ResultSetSourceInfo.java
@@ -3,7 +3,7 @@ package com.microsoft.azure.kusto.ingest.source;
 import java.sql.ResultSet;
 import java.util.UUID;
 
-public class ResultSetSourceInfo extends SourceInfo {
+public class ResultSetSourceInfo extends AbstractSourceInfo {
 
     private ResultSet resultSet;
